@@ -4,29 +4,27 @@
   "description": "",
   "main": "server.js",
   "scripts": {
-    "dev": "SET NODE_ENV=development & node server.js",
+    "dev": "SET NODE_ENV=development & node index.js",
     "build": "next build",
-    "start": "SET NODE_ENV=production & node server.js",
+    "start": "SET NODE_ENV=production & node index.js",
     "test": "jest",
-    "testwatch": "jest --watchAll"
+    "test:watch": "jest --watchAll",
+    "test:coverage": "jest --coverage"
   },
   "author": "",
   "license": "MIT",
   "dependencies": {
     "es6-promise": "^4.2.5",
     "express": "^4.16.3",
-    "i18next": "11.9.0",
-    "i18next-browser-languagedetector": "2.2.3",
-    "i18next-express-middleware": "1.4.0",
-    "i18next-node-fs-backend": "^2.1.3",
-    "i18next-xhr-backend": "1.5.1",
+    "i18next": "^19.1.0",
     "isomorphic-unfetch": "^3.0.0",
-    "next": "^7.0.1",
+    "next": "^9.2.1",
+    "next-i18next": "^4.2.0",
     "next-redux-saga": "^3.0.0",
     "next-redux-wrapper": "^2.0.0",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2",
-    "react-i18next": "7.13.0",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "react-i18next": "^11.3.2",
     "react-redux": "^5.0.7",
     "redux": "^4.0.0",
     "redux-devtools-extension": "^2.13.5",
