@@ -2,10 +2,10 @@ const Breadcrumb = () => (
   <section className="salary-breadcrumb-section" aria-label="breadcrumb">
     <ol className="breadcrumb">
       <li className="breadcrumb-item">
-        <a>{t => <a href="/">{t(900)}</a>}</a>
+        <a>ABC</a>
       </li>
       <li className="breadcrumb-item">
-        <a>{t => <a href="/salary/">{t(396304)}</a>}</a>
+        <a>XYZ</a>
       </li>
       <li className="breadcrumb-item active" aria-current="page">
         <span className="salaryJobTitle3">Title</span>
