@@ -1,54 +1,12 @@
-import actionTypes from "../actions/action-type";
+import { combineReducers } from "redux";
+import index_temp from "./index_temp";
+import homeReducer from "./homeReducer";
+import salaryReducer from "./salaryReducer";
 
-export const exampleInitialState = {
-  count: 0,
-  error: false,
-  lastUpdate: 0,
-  light: false,
-  placeholderData: null
-};
+const allReducers = combineReducers({
+  index_temp,
+  homeReducer,
+  salaryReducer
+});
 
-function reducer(state = exampleInitialState, action) {
-  switch (action.type) {
-    case actionTypes.FAILURE:
-      return {
-        ...state,
-        ...{ error: action.error }
-      };
-
-    case actionTypes.INCREMENT:
-      return {
-        ...state,
-        ...{ count: state.count + 1 }
-      };
-
-    case actionTypes.DECREMENT:
-      return {
-        ...state,
-        ...{ count: state.count - 1 }
-      };
-
-    case actionTypes.RESET:
-      return {
-        ...state,
-        ...{ count: exampleInitialState.count }
-      };
-
-    case actionTypes.LOAD_DATA_SUCCESS:
-      return {
-        ...state,
-        ...{ placeholderData: action.data }
-      };
-
-    case actionTypes.TICK_CLOCK:
-      return {
-        ...state,
-        ...{ lastUpdate: action.ts, light: !!action.light }
-      };
-
-    default:
-      return state;
-  }
-}
-
-export default reducer;
+export default allReducers;
