@@ -3,7 +3,12 @@ import { all, call, put, take, takeLatest } from "redux-saga/effects";
 import es6promise from "es6-promise";
 import "isomorphic-unfetch";
 
-import { failure, loadDataSuccess, tickClock } from "../actions";
+import {
+  failure,
+  loadDataSuccess,
+  tickClock,
+  loadSalarySuccess
+} from "../actions";
 import actionTypes from "../actions/action-type";
 
 es6promise.polyfill();
@@ -26,10 +31,22 @@ function* loadDataSaga() {
   }
 }
 
+function* loadSalaryData() {
+  try {
+    const res = yield fetch(
+      "http://localhost:5009/details?channel=en-US&Occupation=Software%20Developer%20/%20Engineer&Location=Boston,%20MA"
+    );
+    const data = yield res.json();
+    yield put(loadSalarySuccess(data));
+  } catch (err) {
+    yield put(failure(err));
+  }
+}
+
 function* rootSaga() {
   yield all([
     call(runClockSaga),
-    takeLatest(actionTypes.LOAD_DATA, loadDataSaga)
+    takeLatest(actionTypes.LOAD_DATA, loadSalaryData)
   ]);
 }
 
