@@ -1,7 +1,7 @@
 import { createStore, applyMiddleware } from "redux";
 import createSagaMiddleware from "redux-saga";
 import logger from "redux-logger";
-import rootReducer, { exampleInitialState } from "../reducer";
+import rootReducer from "../reducer";
 import rootSaga from "../saga";
 
 const sagaMiddleware = createSagaMiddleware();
@@ -14,7 +14,7 @@ const bindMiddleware = middleware => {
   return applyMiddleware(...middleware);
 };
 
-function configureStore(initialState = exampleInitialState) {
+function configureStore(initialState = []) {
   const store = createStore(
     rootReducer,
     initialState,
