@@ -1,19 +1,20 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
 import { compose } from "redux";
-import { increment } from "../app/actions";
-import { withI18next } from "../app/lib/withI18next";
+import { increment, loadData } from "../app/actions";
+
 export class componentName extends Component {
   Increament = () => {
     this.props.Increament();
   };
+
   render() {
     console.log("test", this.props);
-    const { t } = this.props;
+
     return (
       <div>
         <p>{this.props.count}</p>
-        <button onClick={this.Increament}>{t("bnt_text")}</button>
+        <button onClick={this.props.load}>button text</button>
       </div>
     );
   }
@@ -27,10 +28,11 @@ const mapDispatchToProps = dispatch => {
   return {
     Increament: () => {
       dispatch(increment());
+    },
+    load: () => {
+      console.log("load");
+      dispatch(loadData());
     }
   };
 };
-export default compose(
-  connect(mapStateToProps, mapDispatchToProps),
-  withI18next(["home"])
-)(componentName);
+export default connect(mapStateToProps, mapDispatchToProps)(componentName);
