@@ -1,65 +1,64 @@
 const express = require("express");
-const path = require("path");
 const next = require("next");
+const nextI18NextMiddleware = require("next-i18next/middleware").default;
+
+const nextI18next = require("./i18n");
 
 const dev = process.env.NODE_ENV !== "production";
+const port = process.env.PORT || 3000;
 const app = next({ dev });
 const handle = app.getRequestHandler();
 
-const i18nextMiddleware = require("i18next-express-middleware");
-const Backend = require("i18next-node-fs-backend");
-const i18n = require("./i18n");
+(async () => {
+  await app.prepare();
+  const server = express();
 
-// init i18next with serverside settings
-// using i18next-express-middleware
-i18n
-  .use(Backend)
-  .use(i18nextMiddleware.LanguageDetector)
-  .init(
-    {
-      fallbackLng: "en",
-      preload: ["en", "de"], // preload all langages
-      ns: ["common", "home", "salary"], // need to preload all the namespaces
-      backend: {
-        loadPath: path.join(__dirname, "/locales/{{lng}}/{{ns}}.json"),
-        addPath: path.join(__dirname, "/locales/{{lng}}/{{ns}}.missing.json")
-      }
-    },
-    () => {
-      // loaded translations we can bootstrap our routes
-      app.prepare().then(() => {
-        const server = express();
+  await nextI18next.initPromise;
+  server.use(nextI18NextMiddleware(nextI18next));
 
-        // enable middleware for i18next
-        server.use(i18nextMiddleware.handle(i18n));
+  //  Custom Routes for the salary application
 
-        // serve locales for client
-        server.use(
-          "/locales",
-          express.static(path.join(__dirname, "/locales"))
-        );
-        //Server Side Support for Clean URLs
-        server.get("/p/:id", (req, res) => {
-          const actualPage = "/post";
-          const queryParams = { title: req.params.id };
-          console.log("req.params.id==>", req.params.id);
-          console.log("rreq.query==>", req.query);
+  server.get("/", (req, res) => {
+    const actualPage = "/";
+    app.render(req, res, actualPage);
+  });
 
-          app.render(req, res, actualPage, queryParams);
-        });
+  server.get("/salary/:position/", (req, res) => {
+    const actualPage = "/salary";
+    const queryParams = {
+      lang: req.params.lang,
+      position: req.params.position
+    };
+    app.render(req, res, actualPage, queryParams);
+  });
 
-        // missing keys
-        server.post(
-          "/locales/add/:lng/:ns",
-          i18nextMiddleware.missingKeyHandler(i18n)
-        );
-        // use next.js
-        server.get("*", (req, res) => handle(req, res));
+  server.get("/salary/:position/:location", (req, res) => {
+    const actualPage = "/salary";
+    const queryParams = {
+      lang: req.params.lang,
+      position: req.params.position,
+      location: req.params.location
+    };
 
-        server.listen(3000, err => {
-          if (err) throw err;
-          console.log("> Ready on http://localhost:3000");
-        });
-      });
-    }
-  );
+    //  some data we might need to use to do the checking of the routes to render the app at first time
+    console.log("original Url :: ", req.headers.host);
+    console.log("original Url :: ", req.originalUrl);
+    console.log("base url :: ", req.baseUrl);
+    console.log("query :: ", req.query);
+    console.log("params :: ", req.params);
+    console.log("query params :: ", req.queryParams);
+
+    console.log("path :: ", req._parsedUrl.path);
+    console.log("path name :: ", req._parsedUrl.pathname);
+    console.log("href :: ", req._parsedUrl.href);
+
+    app.render(req, res, actualPage, queryParams);
+  });
+
+  server.get("*", (req, res) => {
+    handle(req, res);
+  });
+
+  await server.listen(port);
+  console.log(`> Ready on http://localhost:${port}`); // eslint-disable-line no-console
+})();
